refactor: reorganize sidebar component imports for better structure
